SDK Refactor and Improvements
