+++ native/ffi/Cargo.toml
@@ -14,13 +14,13 @@ crate-type = ["rlib"]
 [dependencies]
 sdk = { path = "../sdk", package = "subsocial-sdk" }
 bytes = "1.0"
-prost = "0.7"
-prost-types = "0.7"
+prost = "0.9.0"
+prost-types = "0.9.0"
 once_cell = "1.7"
-allo-isolate = { version = "0.1.8-beta", features = ["catch-unwind"] }
+allo-isolate = { version = "0.1.10", features = ["catch-unwind"] }
 # Runtime
 async-std = { version = "1.8", features = [] }
 
 [build-dependencies]
-cbindgen = "0.19"
-prost-build = "0.7"
+cbindgen = "0.20.0"
+prost-build = "0.9.0"
