+++ native/ffi/src/pallet/mod.rs
@@ -1,0 +1,19 @@
+pub(crate) mod system;
+
+pub(crate) mod spaces;
+pub(crate) mod spaces_xt;
+
+pub(crate) mod posts;
+pub(crate) mod posts_xt;
+
+pub(crate) mod reactions;
+pub(crate) mod reactions_xt;
+
+pub(crate) mod profiles;
+pub(crate) mod profiles_xt;
+
+pub(crate) mod space_follows;
+pub(crate) mod space_follows_xt;
+
+pub(crate) mod profile_follows;
+pub(crate) mod profile_follows_xt;
